CVS-ID: src/ChangeLog 1.3202
CVS-ID: src/rcscmds.c 1.71

+++ src/rcscmds.c
@@ -76,7 +76,7 @@ static void RCS_output_diff_options (int, char * const *, const char *,
 
 static char **call_diff_argv;
 static int call_diff_argc;
-static int call_diff_argc_allocated;
+static size_t call_diff_argc_allocated;
 
 static int call_diff (const char *out);
 static int call_diff3 (char *out);
